exercise 4.21

+++ part4/controllers/blogs.js
@@ -4,13 +4,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization .startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -59,8 +52,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({error:'token invalid'})
+  }
+  const user = await User.findById(decodedToken.id) 
+  const blogToBeDeleted = await Blog.findById(request.params.id)
+
+  if(!blogToBeDeleted.id){
+    return response.status(401).json({error: 'blog not found'})
+  }
+
+  if(blogToBeDeleted.user.toString() === user.id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }else{
+    return response.status(403).json({ error: 'The user was not the owner of the blog' })
+  }
 })
 
 module.exports = blogsRouter
